Replace old LARA Plugin API with the compiled TypeScript version
[#165062314]
