+++ app/assets/javascripts/runtime.js
@@ -34,7 +34,7 @@
 //= require modal-dialog
 //= require plugins
 //= require sidebar
-//= require lara-api
+//= require lara-plugin-api
 //= require common
 //= require check-answer
 //= require scroll-interactive-mod
